rename `app` directory to `store`
