+++ frontend/src/common/test-utils.tsx
@@ -13,11 +13,11 @@ import { MemoryRouterProvider } from "next-router-mock/MemoryRouterProvider";
 import React, { PropsWithChildren } from "react";
 import { Provider } from "react-redux";
 
-import { RootState, setupStore } from "@/app/store";
 import { localBackendURL } from "@/common/test-constants";
 import { Faces } from "@/common/types";
 import { setURL } from "@/features/backend/backendSlice";
 import { LayoutWithoutReduxProvider } from "@/features/ui/Layout";
+import { RootState, setupStore } from "@/store/store";
 
 //# region redux test setup
