+++ frontend/src/features/modals/modalsSlice.ts
@@ -1,6 +1,5 @@
 import { PayloadAction } from "@reduxjs/toolkit";
 
-import { RootState } from "@/app/store";
 import {
   CardDocument,
   createAppSlice,
@@ -8,6 +7,7 @@ import {
   ModalsState,
   Slots,
 } from "@/common/types";
+import { RootState } from "@/store/store";
 
 //# region slice configuration
