+++ frontend/src/features/search/sourceDocumentsSlice.ts
@@ -4,14 +4,14 @@
 
 import { createSelector } from "@reduxjs/toolkit";
 
-import { APIGetSources } from "@/app/api";
-import { AppDispatch, RootState } from "@/app/store";
 import {
   createAppAsyncThunk,
   createAppSlice,
   SourceDocumentsState,
 } from "@/common/types";
 import { setNotification } from "@/features/toasts/toastsSlice";
+import { APIGetSources } from "@/store/api";
+import { AppDispatch, RootState } from "@/store/store";
 
 //# region async thunk
